backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[argon2]
pydantic
PyJWT
